+++ common/src/main/java/com/mrpgc_skill_tree/effect/RemoveHarmfulEffectsWhenAppliedStatusEffect.java
@@ -1,0 +1,19 @@
+package com.mrpgc_skill_tree.effect;
+
+import net.minecraft.entity.LivingEntity;
+import net.minecraft.entity.effect.StatusEffect;
+import net.minecraft.entity.effect.StatusEffectCategory;
+
+import static net.more_rpg_classes.util.CustomMethods.clearNegativeEffects;
+
+public class RemoveHarmfulEffectsWhenAppliedStatusEffect extends StatusEffect {
+    protected RemoveHarmfulEffectsWhenAppliedStatusEffect(StatusEffectCategory category, int color) {
+        super(category, color);
+
+    }
+    public void onApplied(LivingEntity entity, int amplifier) {
+        super.onApplied(entity, amplifier);
+        clearNegativeEffects(entity,false);
+    }
+
+}